tools modularization
